requirements.txt
# requirements.txt
tkinter
Pillow==9.5.0
qrcode[pil]==7.4.2
pyzbar==0.1.9
opencv-python==4.8.1.78
keyboard==0.13.5
reportlab==4.0.4
pandas==2.0.3
matplotlib==3.7.2
pywin32==306
sqlite3